+++ app/src/main/java/com/yei/dev/weather_app/ui/theme/Color.kt
@@ -7,5 +7,7 @@ val PurpleGrey80 = Color(0xFFCCC2DC)
 val Pink80 = Color(0xFFEFB8C8)
 
 val Purple40 = Color(0xFF6650a4)
+val Purple700 = Color(0xFF3700B3)
+
 val PurpleGrey40 = Color(0xFF625b71)
 val Pink40 = Color(0xFF7D5260)