+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## V 1.0.0
 ## Added
 - [TASK-001] Initial configuration and setup of the project structure.
+- [HU-001] The presentation screen display was implemented